client: Move vtgateconn plugins into the Vitess Go SQL driver package.
Before this change, users had to explicitly import their preferred vtgateconn implementation (in most cases gRPC).

With this change, this is no longer necessary because the driver will import all available implementations automatically.

This is also more consistent with the remaining code:
- we import different implementations at the binary level (in go/cmd/*)
- the Go SQL Driver is a library and therefore at the same level as a binary
